Skip the halfway implementation using consul
The idea was, to let the cluster form at least through (only) consul.
We could have "numbers-seed.service.consul" to find the seed node. Keeping
the IP in the `netty.tcp.hostname` results in

> dropping message [class akka.actor.ActorSelectionMessage] for non-local
> recipient [Actor[akka.tcp://number-actors@numbers-seed.service.consul:2551/]]
> arriving at [akka.tcp://number-actors@seed.service.consul:2551] inbound
> addresses are [akka.tcp://number-actors@10.0.59.249:2551]

So using the service discovery host name for the seed itself results in a
chicken-egg problem immediately at startup:

> Exception in thread "main" java.net.UnknownHostException: numbers-seed.service.consul: Name or service not known

So consul establishes the host name after/while the service is starting. Pity.

Further, use ZIP artifacts instead of .tar.gz as it says "failed to download".
No clue, what's going on there.